+++ src/main/java/competition/auto_pathplanner/PodiumMidCommandGroup.java
@@ -1,22 +1,20 @@
-package competition.auto;
+package competition.auto_pathplanner;
 
 import competition.commandgroups.FireFromSubwooferCommandGroup;
 import competition.subsystems.pose.PoseSubsystem;
 import competition.subsystems.shooter.commands.ContinuouslyWarmUpForSpeakerCommand;
-import competition.subsystems.shooter.commands.WarmUpShooterRPMCommand;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 
 import javax.inject.Inject;
-import javax.inject.Provider;
 
-public class Far4NoteCommandGroup extends SequentialCommandGroup {
+public class PodiumMidCommandGroup extends SequentialCommandGroup {
 
     @Inject
-    public Far4NoteCommandGroup(PoseSubsystem pose,
-                                FireFromSubwooferCommandGroup firePreload,
-                                Fast4NoteFarCommand fast4NoteFarCommand,
-                                ContinuouslyWarmUpForSpeakerCommand continuouslyWarmUpForSpeakerCommand) {
+    public PodiumMidCommandGroup(PoseSubsystem pose,
+                                 FireFromSubwooferCommandGroup firePreload,
+                                 PodiumMidCommand podiumMidCommand,
+                                 ContinuouslyWarmUpForSpeakerCommand continuouslyWarmUpForSpeakerCommand) {
 
         var startInFrontOfSpeaker = pose.createSetPositionCommand(
                 () -> PoseSubsystem.convertBlueToRedIfNeeded(PoseSubsystem.BlueSubwooferMiddleScoringLocation));
@@ -24,6 +22,6 @@ public class Far4NoteCommandGroup extends SequentialCommandGroup {
 
         this.addCommands(firePreload);
 
-        this.addCommands(Commands.parallel(fast4NoteFarCommand, continuouslyWarmUpForSpeakerCommand));
+        this.addCommands(Commands.parallel(podiumMidCommand, continuouslyWarmUpForSpeakerCommand));
     }
 }
