+++ src/main/java/competition/auto_pathplanner/RobotContainer.java
@@ -2,21 +2,15 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package competition.auto;
+package competition.auto_pathplanner;
 
 import com.pathplanner.lib.auto.AutoBuilder;
 import com.pathplanner.lib.auto.NamedCommands;
 import com.pathplanner.lib.commands.PathPlannerAuto;
-import com.pathplanner.lib.path.PathPlannerPath;
 
-import competition.commandgroups.FireNoteCommandGroup;
-import competition.subsystems.arm.commands.ContinuouslyPointArmAtSpeakerCommand;
-import competition.subsystems.collector.commands.IntakeCollectorCommand;
 import competition.subsystems.collector.commands.IntakeUntilNoteCollectedCommand;
-import competition.subsystems.collector.commands.StopCollectorCommand;
 import competition.subsystems.drive.DriveSubsystem;
 import competition.subsystems.pose.PoseSubsystem;
-import competition.subsystems.shooter.commands.ContinuouslyWarmUpForSpeakerCommand;
 import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
